Destructure matches during type inference.
When we have an expression (MatchX (Var v)) and v is a type variable we
are solving for, we can immediately bind v to the appropriate
constructor, allowing the zonking process to eliminate the match. This
makes it possible to solve through various implicit arguments.

In order to support this, in some cases we need to introduce new names
during unification, so the type is now restricted to UInferM.