Infer top-level implicit args.
When given an un-annotated expression, try to infer implicit type
arguments for any unsolved inference variables that appear in the
type of the expression.

(This makes a few of the tests fail, but I don't think it's a bug
in this change. It seems that there's some failure whenever
you use `buildScoped`/`emitBlock` around a computation with a
local `Fin n` type.)