Add value indexers for records.
There are two meaningful ways to index a record: either slicing
but keeping the axis, or slicing away the axis by using a concrete
value. This makes it possible to slice away axes with a different
syntax, using = instead of :.

This requires changing some of the labeled-items helpers to support
the more elaborate syntax forms.