Run synthesis as part of type inference.
This commit modifies type inference to run dictionary synthesis as a part of
constraint solving, instead of as a separate pass. This has a few advantages:
- Dictionaries are treated just like any other inference variable, and in
  particular show up as names. Eventually, this should enable using `ProjectElt`
  on a dictionary.
- We no longer need to traverse the entire expression to replace ClassDictHoles,
  instead we can just `subst` as usual.

Since variables are bound to atoms, but constructing a dictionary may involve
computation, the result of dictionary synthesis is now a "thunk" `Unit -> ty`,
which is then immediately applied to obtain the dictionary.

This commit also reorganizes some of Inference.hs, and explicitly tracks the
additional class dictionary constraints generated by each instance, instead of
recursively solving for dependencies while solving for the original instance.
Hopefully this leads to clearer errors.

There are still some bugs that need to be fixed, but the simplest test cases
appear to pass.