Closure-convert local class dictionaries to avoid rename errors.
Remembering a scope and trying to synthesize an expression that uses the same
names in scope causes problems, because we may have renamed the names in the
containing expression during substitution. Instead, we can explicitly mark
all of the possible class dictionaries we might use and pass them as inputs
into the dictionary-computing thunk.

This doesn't completely solve the problem, because the TYPES of the expression
might still depend on local arguments that have been renamed. This might be a
larger problem, because the type variables might reference local variables.
Perhaps an interaction with `solveLocal` would solve this problem.