Fix ClassArrow-bound dict inference and application substitution.
ClassArrow-bound dict inference: The new synthesis machinery for superclasses
didn't propagate the ClassArrow-bound arguments to second-level constraints
(i.e.new constraints that arise while building instance dictionaries). This
change makes sure that we look at the local record-of-class-dicts introduced by
second-level constraints in addition to looking at individually bound class
dicts.

Application substitution: Synthesis introduces many immediately-applied lambdas
that show up as visual noise. This adds a quick beta-reduction pass to inline
the results. This also fixes a problem with @noinline: without running a
traversal during synthesis, sometimes the decorator is applied to a Var atom
instead of a Lam atoml, but now we are running a traversal again so the problem
doesn't appear. (This may or may not be indicative of a different bug in
noinline, but it isn't a blocker for now.)