Run synthesis as the last step of inference.
This change allows inference variables to persist while dictionaries are
being synthesized. In particular, ClassDictHole now takes an additional
argument, which will be a type variable that we can constrain. Then, during
synthesis, we check to ensure that the synthesized result is compatible with
the type variable constraints we found.

For now, we simply introduce new type variables and then set them to their
appropriate values.