Allow unification of projected ClassDictHoles.
Associated types are types that exist inside instance dictionaries. Thus, they
can be exposed by running `getProjection` on an instance dictionary. However,
we do not yet know what the dictionary will be until we synthesize it.

To make this work, whenver we project a ClassDictHole, we instead project
its placeholder object. This placeholder object is initialized as a free
inference variable, and it can then be refined using unification.

Later, when we do synthesis, we need to ensure that all of those projections
were well-typed. To do this, we try to inline the class dictionary as an atom,
and then unify that atom with the original placeholder. If this succeeds, we
know that we have synthesized the correct thing, and all of the projected types
will be correct (after zonking).

This also requires:
- unifying projections of type variables by expanding them out to a larger
  object,
- unifying DataCon atoms with each other.